workshop_api_3/pythonProject/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
pydantic
motor  # Para trabalhar com MongoDB de forma assíncrona
